requirements.txt
Pillow
matplotlib
seaborn

pytesseract
